feat(navbar): implement Profile picture on navbar
- refactor useAuthStore.ts
