+++ next.config.ts
@@ -2,6 +2,9 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   /* config options here */
+  images: {
+    domains: ['lh3.googleusercontent.com'], // 구글 이미지 도메인 추가
+  },
 };
 
 export default nextConfig;
